+++ Changelog.md
@@ -1,5 +1,40 @@
 # Changelog
 
+**1.11.2**
+  - bumped to runner 1.11.2 and sameersbn/ubuntu:14.04.20170228
+
+**1.11.0**
+  - added RUNNER_AUTOUNREGISTER option
+
+**1.10.4-2**
+  - added RUNNER_AUTOUNREGISTER option
+
+**1.10.4**
+  - gitlab-ci-multi-runner: upgrade to 1.10.4
+
+**1.10.2**
+  - gitlab-ci-multi-runner: upgrade to 1.10.2
+
+**1.9.3**
+  - gitlab-ci-multi-runner: upgrade to 1.9.3
+  - added support of RUNNER_OUTPUT_LIMIT 
+
+**1.9.0-1**
+  - fix restart of container
+
+**1.9.0**
+ - gitlab-ci-multi-runner: upgrade to 1.9.0
+ - added concurrent jobs support
+
+**1.8.0**
+ - gitlab-ci-multi-runner: upgrade to 1.8.0
+
+**v1.7.0**
+- gitlab-ci-multi-runner: upgrade to v1.7.0
+
+**1.1.4-6**
+ - added docker executor
+
 **1.1.4**
  - gitlab-ci-multi-runner: upgrade to 1.1.4